Added PR templates
Added specific templates for bugfixes, enhancements, documentation, and new features.

The user is required to attest that they have created relevant tests if applicable and have run these tests and existing tests on all supported backends.

Added a basic template for PRs.

The user is required to explain the changes they propose and their benefits (and drawbacks, if applicable).

There is a generic template as well, though its use is discouraged.

---------

Co-authored-by: Slaven Peles <peless@ornl.gov>