Added feature: reusing transpose allocation (#246)
Added the ability to reuse a transpose allocation if a transpose buffer has already been allocated for a given matrix

This feature fits nicely with the working thesis of ReSolve, that memory should not be allocated unnecessarily, and reused between in subsequent linear iterations in the solution of the same nonlinear system.
This is crucial for HyKKT, which requires explicit access to the transpose of some matrices, and may be useful for other methods.