Prohibit sloppy syncing (#284)
Prohibit synchronizing memory space if it is already updated. Only memory space that is out of date can be synchronized with the updated space. For example, if host is up to date and device is out of date, then function `syncData(DEVICE);` will copy data from the host to the device. If the device is up-to-date (i.e. flag `gpu_updated_ == true`), Re::Solve will return an error. The user is required to keep track of which memory space is current.



---------

Co-authored-by: shakedregev <35384901+shakedregev@users.noreply.github.com>